fix: resolve clippy errors across all workspace crates (Refs GH-480)
Allow unwrap()/float_cmp in test/example/bench code across cbtop,
trueno-fft, trueno-image, trueno-quant, trueno-solve, trueno-sparse,
trueno-tensor, trueno-explain, trueno-ptx-debug, trueno-cuda-edge,
trueno-rand, and xtask. Fix raw string hashes, doc backticks, unused
imports, deprecated functions, and misc lints.

+++ trueno-cuda-edge/examples/quant_oracle_demo.rs
@@ -35,7 +35,7 @@ fn demo_format_tolerances() {
 }
 
 /// Format a boundary value for display.
-fn format_boundary_value(v: &f32) -> String {
+fn format_boundary_value(v: &f64) -> String {
     if v.is_nan() {
         "NaN".to_string()
     } else if v.is_infinite() {
